+++ resources/release-notes.md
@@ -2,11 +2,13 @@
 
 - (All) Add shortcut to open app from toolbar (quick access)
 - (All) Fix errors when closing app during mission selection
+- (IMAP) Replace `error` with `uialert` for CPT view when mode or range patterns are incorrect
 
 ## Software
 
 - (All) Add `mag.mixin.TimeSeriesOperationSupport` for simplified operations support
 - (All) Fix issues with colors and legends for multiple stackedplots
+- (IMAP) Fix cropping of events when no events are loaded
 - (IMAP) Check there is data before enabling auto-compression event in `mag.imap.view.Field`
 
 ## Package